# src/main/resources/endpoints/swagger_code_gen_project_meetime_backend.yaml
swagger: '2.0'
info:
  version: '0.0.1'
  title: Project-Meetime-Backend
  description: Expose all endpoints of project MEETIME 2025
  contact:
    name: Leandro
    email: leandroluz201616@gmail.com
host: localhost:8080
basePath: /api/v1
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /webhook/contact-creation:
    post:
      tags:
        - Webhooks
      summary: Handle HubSpot contact creation webhook
      description: Processes events of type "contact.creation" sent by HubSpot webhook.
      operationId: handleWebhook
      parameters:
        - in: header
          name: X-HubSpot-Signature
          type: string
          required: false
          description: Signature Hubspot
        - in: header
          name: X-HubSpot-Signature-Version
          type: string
          required: false
          description: Signature Hubspot
        - in: body
          name: WebhookEvent
          description: Webhook event data
          required: true
          schema:
            $ref: '#/definitions/WebhookHubspotCreateContactRequest'
      responses:
        '200':
          description: Event processed successfully
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/Error'
        '401':
          description: Authorization Failed
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/Error'
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: >
      Insert the JWT token in the format: "Bearer {token}"
security:
  - BearerAuth: []
definitions:
...
  WebhookHubspotCreateContactRequest:
    type: array
    items:
      $ref: '#/definitions/WebhookHubspotCreateContact'
  WebhookHubspotCreateContact:
    type: object
    properties:
      objectId:
        type: integer
        format: int64
      changeSource:
        type: string
      eventId:
        type: integer
        format: int64
      subscriptionId:
        type: integer
      portalId:
        type: integer
      appId:
        type: integer
      occurredAt:
        type: integer
        format: int64
      attemptNumber:
        type: integer
      subscriptionType:
        type: string
      changeFlag:
        type: string
      sourceId:
        type: string
  Error:
    type: array
    items:
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string